requirements.txt
aniso8601==7.0.0
anyascii==0.3.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-debug-toolbar==3.2.4
django-filter==21.1
django-graphiql-debug-toolbar==0.2.0
django-graphql-jwt==0.3.4
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
html5lib==1.1
idna==3.3
l18n==2021.3
mysqlclient==2.1.0
openpyxl==3.0.9
Pillow==8.4.0
promise==2.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
text-unidecode==1.3
urllib3==1.26.7
wagtail==2.15.1
wagtailmedia==0.8.0
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
